Add files via upload
this is a program for the implementation of Doubly Linked List. the structure of the List is such that it contains 2 classes.
1) one that defines the structure of the node i.e data, next and previous
2) defines the Linked List's structure and contains the add and print methods.
The add method is unique as usually there are 3 different methods to add elements -- at the start -- end -- middle but this is a combination of all 3 methods in 1.